major refactoring of game loop
